requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-cloud-storage>=2.10.0
python-multipart>=0.0.6
alembic>=1.12.0
